Refactor models module into it's own folder
The idea is the same, I'll only expose very specific use cases. But
it was getting clautrophic in the file so there we go.
